Return error 500 when gameq fails and correct the type of the score field in the rankings table

+++ database/migrations/2023_07_16_202614_create_player_rankings_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('player_rankings', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedBigInteger('score');
+            $table->integer('score');
             $table->unsignedBigInteger('time');
             $table->unsignedBigInteger('server_id');
